+++ instrumentation/quartz-2.0/library/src/main/java/io/opentelemetry/instrumentation/quartz/v2_0/QuartzTelemetryBuilder.java
@@ -26,7 +26,7 @@ public final class QuartzTelemetryBuilder {
   private final OpenTelemetry openTelemetry;
   private final List<AttributesExtractor<? super JobExecutionContext, ? super Void>>
       additionalExtractors = new ArrayList<>();
-  private boolean captureExperimentalSpanAttributes;
+  private boolean emitExperimentalTelemetry;
   private Function<
           SpanNameExtractor<JobExecutionContext>,
           ? extends SpanNameExtractor<? super JobExecutionContext>>
@@ -48,15 +48,26 @@ public final class QuartzTelemetryBuilder {
   }
 
   /**
-   * Sets whether experimental attributes should be set to spans. These attributes may be changed or
-   * removed in the future, so only enable this if you know you do not require attributes filled by
-   * this instrumentation to be stable across versions
+   * Sets whether experimental telemetry should be emitted. This telemetry may be changed or removed
+   * in the future, so only enable this if you know you do not require telemetry emitted by this
+   * instrumentation to be stable across versions.
    */
+  @CanIgnoreReturnValue
+  public QuartzTelemetryBuilder setEmitExperimentalTelemetry(boolean emitExperimentalTelemetry) {
+    this.emitExperimentalTelemetry = emitExperimentalTelemetry;
+    return this;
+  }
+
+  /**
+   * Sets whether experimental attributes should be set to spans.
+   *
+   * @deprecated Use {@link #setEmitExperimentalTelemetry(boolean)} instead.
+   */
+  @Deprecated
   @CanIgnoreReturnValue
   public QuartzTelemetryBuilder setCaptureExperimentalSpanAttributes(
       boolean captureExperimentalSpanAttributes) {
-    this.captureExperimentalSpanAttributes = captureExperimentalSpanAttributes;
-    return this;
+    return setEmitExperimentalTelemetry(captureExperimentalSpanAttributes);
   }
 
   /**
@@ -83,7 +94,7 @@ public final class QuartzTelemetryBuilder {
     InstrumenterBuilder<JobExecutionContext, Void> instrumenter =
         Instrumenter.builder(openTelemetry, INSTRUMENTATION_NAME, spanNameExtractor);
 
-    if (captureExperimentalSpanAttributes) {
+    if (emitExperimentalTelemetry) {
       instrumenter.addAttributesExtractor(
           AttributesExtractor.constant(AttributeKey.stringKey("job.system"), "quartz"));
       instrumenter.addAttributesExtractor(new SchedulerNameAttributesExtractor());
@@ -98,6 +109,6 @@ public final class QuartzTelemetryBuilder {
     return new QuartzTelemetry(
         new TracingJobListener(instrumenter.buildInstrumenter()),
         eventLogger,
-        captureExperimentalSpanAttributes);
+        emitExperimentalTelemetry);
   }
 }